sim-central-ssc01/packs/sim_meraki_pact/actions/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.2.0
idna==3.3
idna-ssl==1.1.0
meraki==1.19.0
multidict==5.2.0
pkg-resources==0.0.0
python-dotenv==0.19.2
requests==2.27.1
typing-extensions==4.1.1
urllib3==1.26.8
yarl==1.7.2
